Second pass at `Local`
