+++ core/src/main/scala/geotrellis/package.scala
@@ -1,1 +1,1 @@
-package object geotrellis extends Local.ToLocalOps with Local1.ToLocal1Ops with LocalInstances
+package object geotrellis extends Local.ToLocalOps with LocalInstances
